# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /webhook:
    post:
      tags:
      - webhook-controller
      operationId: handleWebhook
      parameters:
      - name: x-hub-signature-256
        in: header
        required: false
        schema:
          type: string
      - name: x-github-event
        in: header
        required: false
        schema:
          type: string
      - name: x-github-delivery
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
  /issues:
    get:
      tags:
      - issue-controller
      operationId: listIssues
      parameters:
      - name: state
        in: query
        required: false
        schema:
          type: string
          default: open
      - name: labels
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      tags:
      - issue-controller
      operationId: createIssue
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{number}/comments:
    post:
      tags:
      - issue-controller
      operationId: createComment
      parameters:
      - name: number
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Comment"
  /issues/{number}:
    get:
      tags:
      - issue-controller
      operationId: getIssue
      parameters:
      - name: number
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Issue"
    patch:
      tags:
      - issue-controller
      operationId: updateIssue
      parameters:
      - name: number
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Issue"
  /events:
    get:
      tags:
      - webhook-controller
      operationId: getRecentEvents
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookEvent"
components:
  schemas:
    IssueCreateRequest:
      required:
      - title
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        labels:
          type: array
          items:
            type: string
    Issue:
      type: object
      properties:
        number:
          type: integer
          format: int32
        title:
          type: string
        body:
          type: string
        state:
          type: string
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        html_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Label:
      type: object
      properties:
        name:
          type: string
    CommentCreateRequest:
      required:
      - body
      type: object
      properties:
        body:
          type: string
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        body:
          type: string
        user:
          $ref: "#/components/schemas/User"
        created_at:
          type: string
          format: date-time
        html_url:
          type: string
    User:
      type: object
      properties:
        login:
          type: string
        id:
          type: integer
          format: int64
    IssueUpdateRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        state:
          type: string
    WebhookEvent:
      type: object
      properties:
        id:
          type: integer
          format: int64
        deliveryId:
          type: string
        event:
          type: string
        action:
          type: string
        issueNumber:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time